network: get ksdata from NM cons, not ifcfg files
We want to work more with NM objects instead of ifcfg files. Ie stop requiring
that each device we are working with in anaconda needs to have ifcfg files (so
creating default ones during initialization). Create a configuration for a
device only when user actually configures the device.  Also allows for better
robustness when NICs are added during instllation.

We still need to work with ifcfgs, mostly to set ONBOOT option on installed
system, but mess with them as little as possible.