network: don't crash on devices with zero MAC address (#1334632)
Resolves: rhbz#1334632

Make handling of ifname= boot option more robust, in this case don't crash if
the device does not have permanent MAC address).

Also add support for multiple ifname= options

And make one log message meaningful.

Details:

When ifname=<DEVNAME>:<MAC> is used we bind DEVNAME to MAC in ifcfg file
if we need to create the ifcfg file in Anaconda.

We need to create the ifcfg in Anaconda if it was not already created in
dracut, ie if the device from ifname is not activated by dracut. (which means
that the ifname= is not applied by dracut, ie the <MAC> device is not renamed
to <DEVAME> though).  When creating ifcfg file for <DEVNAME> we check that
<MAC> from ifname= binding corresponds to actual mac address of the device and
if so we bind the <DEVNAME> to the <MAC> by writing HWADDR value (additionally
to DEVICE) into ifcfg file.