Configure the installation system
For now, the installation system can be specified in the Anaconda
configuration file. The supported types of the system are BOOT_ISO,
LIVE_OS and UNKNOWN. The type will determine what the installer
is allowed to do with the current system.

This is a temporary solution. The type will be replaced with a list
of rules. The rules will be modified by configuration snippets, that
will be installed by packages specific for each type of the system.