network: fix dumping of generic "Wired connection" created in initramfs (#1727904)
With NM in initramfs, for ip=dhcp configuration (or no configuration at all)
there is a generic "Wired connection" created that is not bound to any device
(by "interface-name" setting which is empty), which can be activated on multiple
devices.

If default NM autoconnections are turned off in the installer environment (eg
for RHEL or Server), the connection can not be replaced by the autoconnections
("Wired connection X") created for the devices after switch-root (and NM
restart) by NM.

In this case when creating/ensuring persistent connection/ifcfg for a device we
shouldn't use the generic "Wired connection" but create a new interface-bound
one.