Do not mark ancestors of device with source or stage2 as protected
Resolves: rhbz#35701

I am not able to find any reason why we started to protect also
ancestors of the devices in this case (ie disk in case of source on a
partition).

The change (marking also ancestors of protected devices as protected in
general) was introduced during modularization of Payload.

commit 93fdbff37171972632118c547a0afddd796ffa76

commit 26de57d311c53d3f1833375399e11e926d43cc90

commit 19703fc2ff6e5dfccc74fb45788a10847fd4452c
commit 71dda920a75be5baad7d5b8a93f4f998701b168e

The patch tries to limit the changes to the case of the issue. From what
I was able to look at it seems that it should be OK not to include
ancestors also in other cases.