edns: implement support for --category
Support only 'global' category for now.

Run 'trust anchor' on the certificate dumped into file.

If --path is defined the certificate is dumped to the path, if not
predefined runtime path is used.

If the trust tool is missing, raise exception unless in pre-install
phase where it is expected.

NOTE: it is possible that in the future the import tool (trust) will
be run with parameter (derived from) --category value